Initial Support for PA series (#14)
First cut of support for PA series

So far, I've followed the pattern of the SA series and supported the PA
series to about the same extent as the SA series is supported. All of
the known command codes can be sent and received by the Client but the
State class can only be used for power and mute so far. (Worth noting
that it is necessary to call update() on the State object before power
or mute will work for the SA or PA series.)

I'm working on further changes to add support for decoding more of the
messages and adapt the State class for it. Will write some ideas in the
Issue that I posted.